db updates, enums instead of strings
