+++ Planarian/Planarian.Model/Shared/RequestThrottleKeyType.cs
@@ -1,0 +1,11 @@
+namespace Planarian.Model.Shared;
+
+public enum RequestThrottleKeyType
+{
+    EndpointRateLimit,
+    FileAccess,
+    LoginIp,
+    LoginEmail,
+    PasswordResetIp,
+    PasswordResetEmail
+}
