+++ Planarian/Planarian.Model/Shared/ThrottleProfile.cs
@@ -1,0 +1,9 @@
+namespace Planarian.Model.Shared;
+
+public enum ThrottleProfile
+{
+    Login,
+    PasswordReset,
+    FileAccess,
+    EndpointRateLimit
+}
