+++ Planarian/Planarian/Modules/Account/Services/AccountService.cs
@@ -19,7 +19,6 @@ using Planarian.Modules.Import.Models;
 using Planarian.Modules.Notifications.Services;
 using Planarian.Modules.Tags.Repositories;
 using Planarian.Shared.Base;
-using Planarian.Shared.Throttling;
 
 namespace Planarian.Modules.Account.Services;
