+++ Planarian/Planarian/Modules/Authentication/Services/AuthenticationService.cs
@@ -6,7 +6,6 @@ using Planarian.Modules.Authentication.Repositories;
 using Planarian.Modules.Users.Repositories;
 using Planarian.Shared.Base;
 using Planarian.Shared.Email.Services;
-using Planarian.Shared.Throttling;
 
 namespace Planarian.Modules.Authentication.Services;
