--- Planarian/Planarian/Shared/Throttling/ThrottleProfile.cs
@@ -1,8 +1,0 @@
-namespace Planarian.Shared.Throttling;
-
-public enum ThrottleProfile
-{
-    Login,
-    PasswordReset,
-    FileAccess
-}
